refactor destroy service and package structure
