--- src/components/content/order/CspLogo.tsx
@@ -1,17 +1,0 @@
-import HuaWeiLogo from '../../../img/Huawei_Logo.png';
-import AlibabaLogo from '../../../img/Alibaba_Logo.png';
-import AzureLogo from '../../../img/Microsoft_Azure_Logo.png';
-import AWSLogo from '../../../img/Aws_Logo.png';
-import OpenStackLogo from '../../../img/OpenStack_Logo.png';
-import GoogleLogo from '../../../img/Google_Logo.png';
-import FlexibleEngineLogo from '../../../img/FlexibleEngine_Logo.png';
-import XpanseLogo from '../../../img/Eclipse_Xpanse_Logo.png';
-
-export { HuaWeiLogo };
-export { AlibabaLogo };
-export { AzureLogo };
-export { AWSLogo };
-export { OpenStackLogo };
-export { GoogleLogo };
-export { FlexibleEngineLogo };
-export { XpanseLogo };
