+++ src/components/content/order/destroy/useDestroyRequestSubmitQuery.ts
@@ -1,0 +1,10 @@
+import { useMutation } from '@tanstack/react-query';
+import { ServiceService } from '../../../../xpanse-api/generated';
+
+export function useDestroyRequestSubmitQuery() {
+    return useMutation({
+        mutationFn: (id: string) => {
+            return ServiceService.destroy(id);
+        },
+    });
+}
