+++ src/components/content/order/migrate/ShowDeploy.tsx
@@ -4,9 +4,9 @@
  */
 
 import { CloudServiceProvider, UserAvailableServiceVo } from '../../../../xpanse-api/generated';
-import { OrderItem } from '../OrderSubmit';
+import { OrderItem } from '../create/OrderSubmit';
 import { DeployParam, getDeployParams, MigrationSteps, ParamOnChangeHandler } from '../formElements/CommonTypes';
-import { ApiDoc } from '../ApiDoc';
+import { ApiDoc } from '../../common/ApiDoc';
 import { Button, Form, Input, Space, Tooltip } from 'antd';
 import { InfoCircleOutlined } from '@ant-design/icons';
 import React, { ChangeEvent, useEffect, useState } from 'react';
