+++ src/components/content/order/orderStatus/OrderSubmitFailed.tsx
@@ -4,9 +4,11 @@
  */
 
 import { MigrateSubmitResult, OrderSubmitResult } from './OrderSubmitResult';
-import { convertStringArrayToUnorderedList } from '../../utils/generateUnorderedList';
-import { ApiError, Response, ServiceDetailVo } from '../../../xpanse-api/generated';
+import { convertStringArrayToUnorderedList } from '../../../utils/generateUnorderedList';
+import { ApiError, Response, ServiceDetailVo } from '../../../../xpanse-api/generated';
 import { StopwatchResult } from 'react-timer-hook';
+import { OperationType } from '../formElements/CommonTypes';
+import React from 'react';
 
 function getOrderSubmissionFailedDisplay(reasons: string[]) {
     return (
@@ -19,8 +21,9 @@ function getOrderSubmissionFailedDisplay(reasons: string[]) {
 export function OrderSubmitFailed(
     error: Error,
     deploymentStatus: ServiceDetailVo.serviceDeploymentState,
-    stopWatch: StopwatchResult
-): JSX.Element {
+    stopWatch: StopwatchResult,
+    operationType: OperationType
+): React.JSX.Element {
     if (error instanceof ApiError && 'details' in error.body) {
         const response: Response = error.body as Response;
         return OrderSubmitResult(
@@ -28,7 +31,8 @@ export function OrderSubmitFailed(
             '-',
             'error',
             deploymentStatus,
-            stopWatch
+            stopWatch,
+            operationType
         );
     }
     return OrderSubmitResult(
@@ -36,11 +40,12 @@ export function OrderSubmitFailed(
         '-',
         'error',
         deploymentStatus,
-        stopWatch
+        stopWatch,
+        operationType
     );
 }
 
-export const MigrateSubmitFailed = (error: Error): JSX.Element => {
+export const MigrateSubmitFailed = (error: Error): React.JSX.Element => {
     if (error instanceof ApiError && 'details' in error.body) {
         const response: Response = error.body as Response;
         return MigrateSubmitResult(getOrderSubmissionFailedDisplay(response.details), '-', 'error');
