+++ src/components/content/order/services/Services.tsx
@@ -4,17 +4,17 @@
  */
 
 import { FormOutlined } from '@ant-design/icons';
-import '../../../styles/service_order.css';
+import '../../../../styles/service_order.css';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { createServicePageRoute } from '../../utils/constants';
+import { createServicePageRoute } from '../../../utils/constants';
 import { Col, Empty, Row } from 'antd';
 import { Badge, Space } from 'antd';
-import { sortVersion } from '../../utils/Sort';
-import { ServiceVo, VersionOclVo } from '../../../xpanse-api/generated';
+import { sortVersion } from '../../../utils/Sort';
+import { ServiceVo, VersionOclVo } from '../../../../xpanse-api/generated';
 import ServicesSkeleton from './ServicesSkeleton';
 import ServicesLoadingError from './ServicesLoadingError';
-import { ServicesAvailableService } from '../../../xpanse-api/generated/services/ServicesAvailableService';
+import { ServicesAvailableService } from '../../../../xpanse-api/generated/services/ServicesAvailableService';
 
 function Services(): JSX.Element {
     const [services, setServices] = useState<{ name: string; content: string; icon: string; latestVersion: string }[]>(
